+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Fix for `other_keywords` to accept the optional type as intended.
 - Fix for inheritance issues when including VCDry::DSL on a parent component.
+- Fix to allow passing `nil` to an optional keyword.
 
 ## [1.0.0] - 2023-04-16